refactor: Customer 테이블에서 customerType 삭제
