+++ admin/src/main/java/kernel360/ckt/admin/ui/dto/request/CustomerUpdateRequest.java
@@ -2,7 +2,6 @@ package kernel360.ckt.admin.ui.dto.request;
 
 public record CustomerUpdateRequest (
     String customerName,
-    String customerType,
     String phoneNumber,
     String licenseNumber,
     String zipCode,
